+++ back/src/main/java/com/tdevilleduc/urthehero/back/controller/PersonController.kt
@@ -1,14 +1,11 @@
 package com.tdevilleduc.urthehero.back.controller
 
-import com.tdevilleduc.urthehero.back.constant.ApplicationConstants
 import com.tdevilleduc.urthehero.back.exceptions.PersonInternalErrorException
 import com.tdevilleduc.urthehero.back.model.Person
 import com.tdevilleduc.urthehero.back.model.PersonDTO
 import com.tdevilleduc.urthehero.back.service.IPersonService
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 import org.slf4j.helpers.MessageFormatter
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
@@ -22,16 +19,12 @@ import javax.servlet.http.HttpServletRequest
 @RestController
 @RequestMapping("/api/person")
 internal class PersonController(private val personService: IPersonService) {
-    val logger: Logger = LoggerFactory.getLogger(PersonController::class.java)
 
     @GetMapping(value = ["/all"], produces = [MediaType.APPLICATION_JSON_VALUE])
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "\${swagger.controller.person.get-all.value}", description = "\${swagger.controller.person.get-all.notes}")
     @ResponseBody
     fun getPersons(request: HttpServletRequest): Callable<ResponseEntity<MutableList<Person>>> = Callable {
-        if (logger.isInfoEnabled) {
-            logger.info(ApplicationConstants.CONTROLLER_CALL_LOG, request.requestURI)
-        }
         ResponseEntity.ok(personService.findAll())
     }
 
@@ -41,9 +34,6 @@ internal class PersonController(private val personService: IPersonService) {
     @ResponseBody
     fun getPersonById(request: HttpServletRequest,
                       @PathVariable personId: Int): Callable<ResponseEntity<Person>> = Callable {
-        if (logger.isInfoEnabled) {
-            logger.info(ApplicationConstants.CONTROLLER_CALL_LOG, request.requestURI)
-        }
         ResponseEntity.ok(personService.findById(personId))
     }
 
@@ -51,9 +41,6 @@ internal class PersonController(private val personService: IPersonService) {
     @ResponseBody
     fun createPerson(request: HttpServletRequest,
                      @RequestBody personDto: PersonDTO): Callable<ResponseEntity<PersonDTO>> = Callable {
-        if (logger.isInfoEnabled) {
-            logger.info(ApplicationConstants.CONTROLLER_CALL_LOG, request.requestURI)
-        }
         if (personService.exists(personDto.id!!)) {
             throw PersonInternalErrorException(MessageFormatter.format("Une personne avec l'identifiant {} existe déjà. Elle ne peut être créée", personDto.id).message)
         }
@@ -64,9 +51,6 @@ internal class PersonController(private val personService: IPersonService) {
     @ResponseBody
     fun updatePerson(request: HttpServletRequest,
                      @RequestBody personDto: PersonDTO): Callable<ResponseEntity<PersonDTO>> = Callable {
-        if (logger.isInfoEnabled) {
-            logger.info(ApplicationConstants.CONTROLLER_CALL_LOG, request.requestURI)
-        }
         Assert.notNull(personDto.id) { throw PersonInternalErrorException("L'identifiant de la personne passée en paramètre ne peut pas être null") }
         ResponseEntity.ok(personService.createOrUpdate(personDto))
     }
@@ -75,9 +59,6 @@ internal class PersonController(private val personService: IPersonService) {
     @ResponseBody
     fun deletePerson(request: HttpServletRequest,
                      @PathVariable personId: Int) {
-        if (logger.isInfoEnabled) {
-            logger.info(ApplicationConstants.CONTROLLER_CALL_LOG, request.requestURI)
-        }
         personService.delete(personId)
     }